Fix catalog with sysl
