+++ pkg/catalog/swaggerui_test.go
@@ -1,0 +1,7 @@
+package catalog
+
+import "testing"
+
+func TestSwaggerUI(t *testing.T) {
+
+}
